+++ test/unit/editors/ied/__snapshots__/do-container.test.snap.js
@@ -1,6 +1,33 @@
 /* @web/test-runner snapshot v1 */
 export const snapshots = {};
 
+snapshots["looks like the latest snapshot with a SDO element."] = 
+`<action-pane
+  icon=""
+  tabindex="0"
+>
+  <abbr slot="action">
+    <mwc-icon-button
+      icon="info"
+      title="someSdo"
+    >
+    </mwc-icon-button>
+  </abbr>
+  <abbr
+    slot="action"
+    title="[iededitor.toggleChildElements]"
+  >
+    <mwc-icon-button-toggle
+      id="toggleButton"
+      officon="keyboard_arrow_down"
+      onicon="keyboard_arrow_up"
+    >
+    </mwc-icon-button-toggle>
+  </abbr>
+</action-pane>
+`;
+/* end snapshot looks like the latest snapshot with a SDO element. */
+
 snapshots["looks like the latest snapshot with a SDO element and child elements are toggled."] = 
 `<action-pane
   icon=""
@@ -34,31 +61,4 @@ snapshots["looks like the latest snapshot with a SDO element and child elements 
 </action-pane>
 `;
 /* end snapshot looks like the latest snapshot with a SDO element and child elements are toggled. */
-
-snapshots["looks like the latest snapshot with a SDO element."] = 
-`<action-pane
-  icon=""
-  tabindex="0"
->
-  <abbr slot="action">
-    <mwc-icon-button
-      icon="info"
-      title="someSdo"
-    >
-    </mwc-icon-button>
-  </abbr>
-  <abbr
-    slot="action"
-    title="[iededitor.toggleChildElements]"
-  >
-    <mwc-icon-button-toggle
-      id="toggleButton"
-      officon="keyboard_arrow_down"
-      onicon="keyboard_arrow_up"
-    >
-    </mwc-icon-button-toggle>
-  </abbr>
-</action-pane>
-`;
-/* end snapshot looks like the latest snapshot with a SDO element. */