+++ test/unit/editors/ied/__snapshots__/ln-container.test.snap.js
@@ -1,7 +1,7 @@
 /* @web/test-runner snapshot v1 */
 export const snapshots = {};
 
-snapshots["ln-container looks like the latest snapshot with a LN0 element."] = 
+snapshots["ln-container looks like the latest snapshot with a LN element."] = 
 `<action-pane tabindex="0">
   <abbr
     slot="action"
@@ -16,7 +16,7 @@ snapshots["ln-container looks like the latest snapshot with a LN0 element."] =
   </abbr>
 </action-pane>
 `;
-/* end snapshot ln-container looks like the latest snapshot with a LN0 element. */
+/* end snapshot ln-container looks like the latest snapshot with a LN element. */
 
 snapshots["ln-container looks like the latest snapshot with a LN0 element and child elements are toggled."] = 
 `<action-pane tabindex="0">